refactor: rename namespaces to be plural
