+++ src/AppModel/NetDaemon.AppModel/Internal/AppFactories/AppFactoryHelper.cs
@@ -1,6 +1,6 @@
 ﻿using System.Reflection;
 
-namespace NetDaemon.AppModel.Internal.AppFactory;
+namespace NetDaemon.AppModel.Internal.AppFactories;
 
 internal static class AppFactoryHelper
 {
