--- src/AppModel/NetDaemon.AppModel/Internal/AppFactoryProvider/IAppFactoryProvider.cs
@@ -1,8 +1,0 @@
-using NetDaemon.AppModel.Internal.AppFactory;
-
-namespace NetDaemon.AppModel.Internal.AppFactoryProvider;
-
-internal interface IAppFactoryProvider
-{
-    IReadOnlyCollection<IAppFactory> GetAppFactories();
-}
