+++ src/AppModel/NetDaemon.AppModel/Internal/AppFactoryProviders/IAppFactoryProvider.cs
@@ -1,0 +1,8 @@
+using NetDaemon.AppModel.Internal.AppFactories;
+
+namespace NetDaemon.AppModel.Internal.AppFactoryProviders;
+
+internal interface IAppFactoryProvider
+{
+    IReadOnlyCollection<IAppFactory> GetAppFactories();
+}
